Add files via upload
Number Challenge is a playful and interactive number game with four exciting levels that challenge your mind while keeping it fun. Built using HTML, CSS, and JavaScript, it showcases creativity and smart use of modern technology. With four engaging levels—guessing, challenges, sequences, and bubble hunts—it blends fun with learning in a refreshing way. Designed to be light, engaging, and enjoyable, it offers a quick challenge for kids and adults alike while highlighting innovative thinking and problem-solving skills. This project highlights the effective use of effectively use AI-powered tools to build, debug, and deploy a complete web application efficiently.